[chore] docs: add list_deployment docs
