+++ featurebyte/common/documentation/gen_ref_pages_docs_builder.py
@@ -551,7 +551,8 @@ def _add_pure_methods_to_doc_groups(
     """
     methods = [
         ("featurebyte.core.timedelta", "to_timedelta"),
-        ("featurebyte.feature_utility", "list_unsaved_features"),
+        ("featurebyte.list_utility", "list_unsaved_features"),
+        ("featurebyte.list_utility", "list_deployments"),
     ]
     for method in methods:
         doc_groups[
